after double

+++ include/lib0.hpp
@@ -15,48 +15,48 @@
 
 
 
-std::vector<float> zeros(int points);
-std::vector<float> vec_from_array(PyArrayObject *array);
-PyArrayObject *array_from_vec(std::vector<float> vec);
-std::vector<float> costhetaFunc(const std::vector<float> &theta);
-std::vector<float> sinthetaFunc(const std::vector<float> &theta);
-std::vector<float> linspace(float start, float end, size_t points);
-std::vector<float> density(float rho_0, float alpha_0, float rho_1, float alpha_1,std::vector<float> r );
+std::vector<double> zeros(int points);
+std::vector<double> vec_from_array(PyArrayObject *array);
+PyArrayObject *array_from_vec(std::vector<double> vec);
+std::vector<double> costhetaFunc(const std::vector<double> &theta);
+std::vector<double> sinthetaFunc(const std::vector<double> &theta);
+std::vector<double> linspace(double start, double end, size_t points);
+std::vector<double> density(double rho_0, double alpha_0, double rho_1, double alpha_1,std::vector<double> r );
 
 
-float get_g_cpu( int ii, int jj, float G, float dv0, const std::vector<float>& r, const std::vector<float>& z,
-                 const std::vector<float>& costheta, const std::vector<float>& sintheta, const std::vector<float>& rho);
+double get_g_cpu( int ii, int jj, double G, double dv0, const std::vector<double>& r, const std::vector<double>& z,
+                 const std::vector<double>& costheta, const std::vector<double>& sintheta, const std::vector<double>& rho);
 
-std::vector<float> get_all_g_impl_cpu(int nr, int nz, float G, float dv0, const std::vector<float>& r, const std::vector<float>& z,
-                                      const std::vector<float>& costheta, const std::vector<float>& sintheta, const std::vector<float>& rho);
+std::vector<double> get_all_g_impl_cpu(int nr, int nz, double G, double dv0, const std::vector<double>& r, const std::vector<double>& z,
+                                      const std::vector<double>& costheta, const std::vector<double>& sintheta, const std::vector<double>& rho);
 
-__global__ void get_all_g_kernel(int nr, int nz, int nr_sampling, int nz_sampling, float G, float dv0,
-                                 const float *r_sampling, const float *z_sampling,
-                                 const float *grid_data, const float *rho,
-                                 int costheta_size, bool radial, float *f_z);
+__global__ void get_all_g_kernel(int nr, int nz, int nr_sampling, int nz_sampling, double G, double dv0,
+                                 const double *r_sampling, const double *z_sampling,
+                                 const double *grid_data, const double *rho,
+                                 int costheta_size, bool radial, double *f_z);
 
-std::vector<float> get_all_g_impl_cuda(float G, float dv0,
-                                       const std::vector<float> &r_sampling, const std::vector<float> &z_sampling,
-                                       const std::vector<float> &r, const std::vector<float> &z,
-                                       const std::vector<float> &costheta, const std::vector<float> &sintheta,
-                                       const std::vector<float> &rho, float redshift = 0.0, bool radial = true);
+std::vector<double> get_all_g_impl_cuda(double G, double dv0,
+                                       const std::vector<double> &r_sampling, const std::vector<double> &z_sampling,
+                                       const std::vector<double> &r, const std::vector<double> &z,
+                                       const std::vector<double> &costheta, const std::vector<double> &sintheta,
+                                       const std::vector<double> &rho, double redshift = 0.0, bool radial = true);
 
-std::vector<float> get_all_g(float G, float dv0,
-                             const std::vector<float> &r_sampling, const std::vector<float> &z_sampling,
-                             const std::vector<float> &r, const std::vector<float> &z,
-                             const std::vector<float> &costheta, const std::vector<float> &sintheta,
-                             const std::vector<float> &rho, float redshift = 0.0, bool radial = true);
+std::vector<double> get_all_g(double G, double dv0,
+                             const std::vector<double> &r_sampling, const std::vector<double> &z_sampling,
+                             const std::vector<double> &r, const std::vector<double> &z,
+                             const std::vector<double> &costheta, const std::vector<double> &sintheta,
+                             const std::vector<double> &rho, double redshift = 0.0, bool radial = true);
 
-double massCalc(float alpha, float rho, float h);
-double massCalcX(float alpha, float rho, float h, float x);
+double massCalc(double alpha, double rho, double h);
+double massCalcX(double alpha, double rho, double h, double x);
 
 
-std::vector<float> calculate_rotational_velocity(float G, float dv0,
-                                                 std::vector<float> r_sampling,
-                                                 const std::vector<float> &r,
-                                                 const std::vector<float> &z,
-                                                 const std::vector<float> &costheta,
-                                                 const std::vector<float> &sintheta,
-                                                 const std::vector<float> &rho);
+std::vector<double> calculate_rotational_velocity(double G, double dv0,
+                                                 std::vector<double> r_sampling,
+                                                 const std::vector<double> &r,
+                                                 const std::vector<double> &z,
+                                                 const std::vector<double> &costheta,
+                                                 const std::vector<double> &sintheta,
+                                                 const std::vector<double> &rho);
